Add encrypter, re-organize code
